+++ src/main/kotlin/com/learnspigot/bot/help/search/SearchHelpCommand.kt
@@ -1,8 +1,8 @@
 package com.learnspigot.bot.help.search
 
 import com.learnspigot.bot.Registry
-import com.learnspigot.bot.Server.replyEphemeral
 import com.learnspigot.bot.util.embed
+import com.learnspigot.bot.util.replyEphemeral
 import gg.flyte.neptune.annotation.Command
 import gg.flyte.neptune.annotation.Description
 import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel
