+++ src/main/kotlin/com/learnspigot/bot/knowledgebase/KnowledgebaseListener.kt
@@ -2,7 +2,7 @@ package com.learnspigot.bot.knowledgebase
 
 import com.learnspigot.bot.Registry
 import com.learnspigot.bot.Server
-import com.learnspigot.bot.Server.isChannel
+import com.learnspigot.bot.util.isChannel
 import net.dv8tion.jda.api.entities.channel.ChannelType
 import net.dv8tion.jda.api.entities.channel.unions.ChannelUnion
 import net.dv8tion.jda.api.events.channel.ChannelCreateEvent
